+++ src/components/settings/UserManagementSection.tsx
@@ -19,24 +19,8 @@ import {
   encodeServerMagicString,
 } from '../../utils/serverMagicString';
 import { shortHostFromServerUrl } from '../../utils/serverDisplayName';
+import { formatLastSeen } from '../../utils/userMgmtHelpers';
 import { UserForm, type UserFormState } from './UserForm';
-
-function formatLastSeen(iso: string | null | undefined, locale: string, neverLabel: string): string {
-  if (!iso) return neverLabel;
-  const t = new Date(iso).getTime();
-  // Navidrome returns "0001-01-01T00:00:00Z" for never-accessed users → guard against bogus epochs.
-  if (!Number.isFinite(t) || t < 1_000_000_000_000) return neverLabel;
-  const diffSec = (t - Date.now()) / 1000;
-  const abs = Math.abs(diffSec);
-  const rtf = new Intl.RelativeTimeFormat(locale, { numeric: 'auto' });
-  if (abs < 60) return rtf.format(Math.round(diffSec), 'second');
-  if (abs < 3600) return rtf.format(Math.round(diffSec / 60), 'minute');
-  if (abs < 86400) return rtf.format(Math.round(diffSec / 3600), 'hour');
-  if (abs < 604800) return rtf.format(Math.round(diffSec / 86400), 'day');
-  if (abs < 2592000) return rtf.format(Math.round(diffSec / 604800), 'week');
-  if (abs < 31536000) return rtf.format(Math.round(diffSec / 2592000), 'month');
-  return rtf.format(Math.round(diffSec / 31536000), 'year');
-}
 
 export function UserManagementSection({
   serverUrl,
